Title: 최소 에너지 비용, Time: 235ms, Memory: 16MB, Status: Wrong Answer - Codetree

+++ 250726/최소 에너지 비용/minimum-energy-cost.java
@@ -17,7 +17,7 @@ public class Main {
 
         long sum = 0;
         for (int i = 0; i < n - 1; i++)
-            sum += (dist[i] * lowCost[i]);
+            sum += (long)(dist[i] * lowCost[i]);
 
         System.out.print(sum);
     }
